requirements.txt
openai>=1.3.6
sqlalchemy
streamlit==1.56.0
pillow>=12.2.0,<13
tiktoken
matplotlib
seaborn
audio_recorder_streamlit==0.0.10
dashscope
pandas>=2.0.0
openpyxl==3.1.2
pdfplumber==0.10.2